# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "SimbaTracking",
        "description": "SimbaTracking"
    },
    "paths": {
...
        "/customer/update-profile": {
            "post": {
                "x-controller-name": "CustomerController",
                "x-operation-name": "UpdateProfile",
                "tags": ["CustomerController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "UpdateProfile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "customer_id": {
                                        "type": "string"
                                    },
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "profile_photo": {
                                        "type": "file"
                                    },
                                    "customer_type": {
                                        "type": "string",
                                        "description": "individual or business"
                                    },
                                    "business_name": {
                                        "type": "string",
                                        "description": "Pass while customer_type=business"
                                    },
                                    "business_mobile_country_code": {
                                        "type": "string",
                                        "description": "Pass while customer_type=business"
                                    },
                                    "business_mobile": {
                                        "type": "string",
                                        "description": "Pass while customer_type=business"
                                    },
                                    "preferred_contact": {
                                        "type": "string",
                                        "description": "Pass while customer_type=business"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "CustomerController.UpdateProfile"
            }
        },
...
        "/driver/check-driver": {
            "post": {
                "x-controller-name": "DriverController",
                "x-operation-name": "CheckDriver",
                "tags": ["DriverController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "CheckDriver",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "instance_id": {
                                        "type": "string"
                                    },
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "mobile_country_code": {
                                        "type": "string"
                                    },
                                    "mobile": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "DriverController.CheckDriver"
            }
        },
        "/driver/get-profile": {
            "get": {
                "x-controller-name": "DriverController",
                "x-operation-name": "GetProfile",
                "tags": ["DriverController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "GetProfile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "parameters": [{
                    "name": "driver_id",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }],
                "operationId": "DriverController.GetProfile"
            }
        },
        "/driver/signup": {
            "post": {
                "x-controller-name": "DriverController",
                "x-operation-name": "SignupDriver",
                "tags": ["DriverController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "SignupDriver",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "instance_id": {
                                        "type": "string"
                                    },
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "mobile_country_code": {
                                        "type": "string"
                                    },
                                    "mobile": {
                                        "type": "string"
                                    },
                                    "licence_number": {
                                        "type": "string"
                                    },
                                    "licence_photo": {
                                        "type": "file"
                                    },
                                    "profile_photo": {
                                        "type": "file"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "dob": {
                                        "type": "string",
                                        "description": "Format YYYY-MM-DD"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "family_mobile_country_code": {
                                        "type": "string"
                                    },
                                    "family_mobile": {
                                        "type": "string"
                                    },
                                    "device_token": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "DriverController.SignupDriver"
            }
        },
         "/driver/update-profile": {
            "post": {
                "x-controller-name": "DriverController",
                "x-operation-name": "UpdateProfile",
                "tags": ["DriverController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "UpdateProfile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "driver_id": {
                                        "type": "string"
                                    },
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "mobile_country_code": {
                                        "type": "string"
                                    },
                                    "mobile": {
                                        "type": "string"
                                    },
                                    "licence_number": {
                                        "type": "string"
                                    },
                                    "licence_photo": {
                                        "type": "file"
                                    },
                                    "profile_photo": {
                                        "type": "file"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "dob": {
                                        "type": "string",
                                        "description": "Format YYYY-MM-DD"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "family_mobile_country_code": {
                                        "type": "string"
                                    },
                                    "family_mobile": {
                                        "type": "string"
                                    },
                                    "device_token": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "DriverController.UpdateProfile"
            }
        },
        "/driver/check-instance-connet": {
            "get": {
                "x-controller-name": "DriverController",
                "x-operation-name": "CheckInstanceConnect",
                "tags": ["DriverController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "CheckInstanceConnect",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "parameters": [{
                    "name": "driver_id",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }],
                "operationId": "DriverController.CheckInstanceConnect"
            }
        },
        "/driver/logout": {
            "post": {
                "x-controller-name": "DriverController",
                "x-operation-name": "Logout",
                "tags": ["DriverController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "Logout",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "driver_id": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "DriverController.Logout"
            }
        },
        "/driver/signin": {
            "post": {
                "x-controller-name": "DriverController",
                "x-operation-name": "SignIn",
                "tags": ["DriverController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "SignIn",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "mobile_country_code": {
                                        "type": "string"
                                    },
                                    "mobile": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "DriverController.SignIn"
            }
        },
...
        "/driver/routes-list": {
            "get": {
                "x-controller-name": "DriverController",
                "x-operation-name": "RoutesList",
                "tags": ["DriverController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "RoutesList",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "parameters": [{
                    "name": "driver_id",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }],
                "operationId": "DriverController.RoutesList"
            }
        },
        "/driver/job-list": {
            "get": {
                "x-controller-name": "DriverController",
                "x-operation-name": "JobList",
                "tags": ["DriverController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "JobList",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "parameters": [{
                    "name": "driver_id",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }],
                "operationId": "DriverController.JobList"
            }
        },
...
        "/driver/history-list": {
            "get": {
                "x-controller-name": "DriverController",
                "x-operation-name": "HistoryList",
                "tags": ["DriverController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "HistoryList",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "parameters": [{
                    "name": "driver_id",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }],
                "operationId": "DriverController.HistoryList"
            }
        }
    },
...
    "servers": [{
        "url": "http://localhost:1330/api"
    }, {
        "url": "http://15.188.211.169:1330/api"
    }]
}